Fixed the stick plot and added PV profile
Fixed the stick plot function, the graph area wasn't being passed as a parameter.
Added the Pseudo-Voigt profile with a ratio where the simulated experimental resolution is used as the Gaussian width and the transition's width is used as the Lorentzian width
Also refactored the profile calls into a single function where we can add more profiles in the future.
Updated documentation.