Fixed some references to make the excel export work for ion spectra
There was some problems with the excel file export where some variables were not referenced in the function call. Made them be available in the variables.py file to be able to reference them in any module. There might be some places where it might not be fully fixed but it works now.