Added some conversion scripts
Added input conversion scripts to format the input rates and spectra files that are generated from the MCDF calculations. The output of these scripts will be the file in the correct format to be read by the simulations program. The orbital to layer notation conversion is not perfect, also because sometimes due to the LS -> jj conversion we can't differentiate some layers, i.e. L2 from L3. I had a better version but of these scripts but I can't find them so we have to be careful to configure the ifs when using for a new electron configuration.