Completed the last commit and added the beam energy filter
Added the filtering option for the beam energy. This is done using the files "z-grounddiagenergy.out" and "z-groundsatenergy.out" for the diagram/auger and satellite respectively. The beam energy is compared to the energy in the file using its low level and is added to be plotted if we have enough energy.
Also fixed various other things due with satellite transitions that I can no longer remember what they where.