Added simulated excitation beam profile
Improved the line selection by using the excitation beam. We now can define a FWHM for a gaussian profile to modulate the intensity of selected transitions. This modulation is performed according to the integral of the gaussian beam profile from the formation energy of the transition we wish to filter, until infinity (using scipy).