Added Satellite intensity from Auger process
The satellite lines intensity now takes into consideration the branching ratio from higher shells Auger transitions to form the transitions being plotted.
For a beam energy of 0 this is not taken into account. This is temporary, for test purposes.
Also added a script to convert auger rates and calculate their branching ratio